+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   ([@nashby](https://github.com/nashby))
 
 ### bug fix
+  * Allow to specify checked and uncked values for boolean input
+  ([@nashby](https://github.com/nashby)).
+  Closes [#643](https://github.com/plataformatec/simple_form/issues/629)
   * Allow to add additional classes only for wrapper.
   ([@nashby](https://github.com/nashby)).
   Closes [#629](https://github.com/plataformatec/simple_form/issues/629)